Move NashResult to utilities.py
