+++ quantecon/game_theory/mclennan_tourky.py
@@ -10,6 +10,7 @@ import numbers
 import numpy as np
 from ..compute_fp import _compute_fixed_point_ig
 from .normal_form_game import pure2mixed
+from .utilities import NashResult
 
 
 def mclennan_tourky(g, init=None, epsilon=1e-3, max_iter=200,
@@ -297,48 +298,3 @@ def _flatten_action_profile(action_profile, indptr):
         out[indptr[i]:indptr[i+1]] = mixed_action
 
     return out
-
-
-class NashResult(dict):
-    """
-    Contain the information about the result of Nash equilibrium
-    computation.
-
-    Attributes
-    ----------
-    NE : tuple(ndarray(float, ndim=1))
-        Computed Nash equilibrium.
-
-    converged : bool
-        Whether the routine has converged.
-
-    num_iter : int
-        Number of iterations.
-
-    max_iter : int
-        Maximum number of iterations.
-
-    init_action_profile : array_like
-        Initial action profile used.
-
-    """
-    # This is sourced from sicpy.optimize.OptimizeResult.
-    def __getattr__(self, name):
-        try:
-            return self[name]
-        except KeyError:
-            raise AttributeError(name)
-
-    __setattr__ = dict.__setitem__
-    __delattr__ = dict.__delitem__
-
-    def __repr__(self):
-        if self.keys():
-            m = max(map(len, list(self.keys()))) + 1
-            return '\n'.join([k.rjust(m) + ': ' + repr(v)
-                              for k, v in sorted(self.items())])
-        else:
-            return self.__class__.__name__ + "()"
-
-    def __dir__(self):
-        return self.keys()
